save hotel and create hotelMapper

+++ city_explorer_back/src/main/java/com/example/city_explorer/services/Impl/RefreshTokenServiceImpl.java
@@ -49,6 +49,9 @@ public class RefreshTokenServiceImpl implements RefreshTokenService {
     @Override
     public void deleteOldToken(User user) {
       var refreshToken= refreshTokenRepository.findByUser(user);
+      if (refreshToken == null){
+          return;
+      }
       refreshTokenRepository.delete(refreshToken);
     }
 }
